requirements.txt
# Core
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12

# Deep Learning
torch>=2.0
torchvision>=0.15

# Machine Learning
scikit-learn>=1.3
scipy>=1.11

# Dimensionality Reduction
umap-learn>=0.5

# Image Processing
Pillow>=10.0

# Utilities
tqdm>=4.65
